move collect types to voiceCollect
