+++ packages/core/src/types/voice.ts
@@ -11,5 +11,6 @@ export * from './voicePlayback'
 export * from './voiceRecording'
 export * from './voiceDetect'
 export * from './voiceTap'
+export * from './voiceCollect'
 
 export type VoiceJSONRPCMethod = VoiceCallJSONRPCMethod
